Changed VoiceTrigger to only write to disk at specific times when explicitly called.
Writing to disk takes a lot of time, so it can interfere with
stim/response timing. Now sounds are saved in RAM until explicitly
written to disk from the experiment script. VT_test_exp.py is updated
to reflect this. (I don't think anyone but me is using this yet, so
it's not back-compatible.)